over over code and clean up things
