+++ code/compose/ui-theme/src/main/AndroidManifest.xml
@@ -1,4 +1,1 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest>
-
-</manifest>
+<manifest package="com.nicholas.rutherford.habit.vibes.quote.api.manager.compose.ui.theme" />
