+++ code/compose/ui-theme/src/main/java/com/nicholas/rutherford/habit/vibes/quote/api/manager/compose/ui/theme/Color.kt
@@ -9,8 +9,6 @@ val SecondaryDark = Color(0xFFB0BEC5)
 val Tertiary = Color(0xFF7CB342)
 val TertiaryDark = Color(0xFFAED581)
 
-val LightBackground = Color(0xFFFAFAFA)
-val DarkBackground = Color(0xFF121212)
 val LightSurface = Color(0xFFFFFFFF)
 val DarkSurface = Color(0xFF1E1E1E)
 val LightOnSurface = Color(0xFF212121)
@@ -20,5 +18,3 @@ val DarkError = Color(0xFFEF9A9A)
 
 val LightScreenBackground = Color(0xFFF5F7FA)
 val DarkScreenBackground = Color(0xFF1A1C1E)
-val LightBottomNavBackground = Color(0xFFE9EEF4)
-val DarkBottomNavBackground = Color(0xFF2A2D31)
