+++ code/navigation/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.nicholas.rutherford.habit.vibes.quote.api.manager.navigation" />
